pyproject.toml
[project]
name = "shortmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "openai>=1.88.0",
    "runware>=0.4.13",
    "moviepy>=1.0.3",
    "numpy>=1.26.4",
    "Pillow>=9.5.0",
]
